Test commit markdown
Trying to build from an incomplete configuration, i.e. `:Cmake` with
errors and then `:make`, set the quickfix list to an error in the
generated Makefile under the build directory. This resulted in an empty
file while trying to follow the error.
Now the `CMakeList.txt` file, which contains the error, will be opened.

The link for `Makefile|xx| recipe for target '...' failed` is still not
working.